pyproject.toml
[tool.poetry]
name = "pythonmyadmin"
version = "0.1.0"
description = "Lightweight GUI for modifying database tables."
authors = ["Todd Birchard <toddbirchard@gmail.com>"]
maintainers = ["Todd Birchard <toddbirchard@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/toddbirchard/pythonmyadmin/"
repository = "https://github.com/toddbirchard/pythonmyadmin/"
documentation = "https://github.com/toddbirchard/pythonmyadmin/"
keywords = [
    "Python",
    "SQL",
    "MySQL",
    "Postgres",
    "Tooling"
]

[tool.poetry.dependencies]
python = "^3.8"
flask = "*"
flask-sqlalchemy = "*"
flask-session = "*"
pymysql = "*"
flask-assets = "*"
dash = "*"
dash-core-components = "*"
dash-html-components = "*"
dash-renderer = "*"
pandas = "*"
wtforms = "*"
redis = "*"
python-dotenv = "*"

[tool.poetry.dev-dependencies]
Pytest = "*"
Lesscpy = "*"
Cssmin = "*"
Jsmin = "*"
black = "^20.8b1"
isort = "^5.6.4"

[tool.poetry.scripts]
run = "wsgi:app"

[tool.poetry.urls]
issues = "https://github.com/toddbirchard/pythonmyadmin/issues"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
